Made the theme compatible with Hugo 0.18.
At least in Hugo 0.18 (I don't know for prior versions), the variables in the
templates are all set to lowercase even if they are not lowercase in the
config file.

For instance, for a configuration as follows:
[Author]
   Handle = "bla"
the variable must be used as .Site.Author.handle and not .Site.Author.Handle.